Extract DPM wave campaign definition service boundary
